+++ SimulationExercise.Core/DTOs/DatabaseDTOs/AverageProvinceDataInsertDTO.cs
@@ -6,14 +6,14 @@ namespace SimulationExercise.Core.DTOs.DatabaseDTOs
     {
         public AverageProvinceDataInsertDTO(string province,
                                             string sensorTypeName,
-                                            float averageValue,
+                                            double averageValue,
                                             Unit unit,
                                             int averageDaysOfMeasure,
                                             bool isExported)
         {
             Province = province;
             SensorTypeName = sensorTypeName;
-            AverageValue = averageValue;
+            AverageValue = (float)averageValue;
             Unit = unit;
             AverageDaysOfMeasure = averageDaysOfMeasure;
             IsExported = isExported;
@@ -21,7 +21,7 @@ namespace SimulationExercise.Core.DTOs.DatabaseDTOs
 
         public string Province { get; }
         public string SensorTypeName { get; }
-        public float AverageValue { get; }
+        public double AverageValue { get; }
         public Unit Unit { get; }
         public int AverageDaysOfMeasure { get; }
         public bool IsExported { get; set; }
